+++ MIDI.cs
@@ -78,7 +78,7 @@ internal class MIDI : Device
         return 0xee;
     }
 
-    public override bool Tick(int ticks, int ignored)
+    public override bool Tick(int ticks, long ignored)
     {
         return false;
     }
